Lesion_1_fixed1

+++ area_of_​​a_triangle.rb
@@ -1,6 +1,6 @@
 puts 'Введите длинну основания треугольника'
-height = gets.to_i
+height = gets.to_f
 puts 'Введите высоту треугольника'
-base = gets.to_i
+base = gets.to_f
 area = 0.5 * (height * base)
 puts "лощадь треугольника = #{area}"
