+++ weight.rb
@@ -1,7 +1,7 @@
 puts 'Здравствуйте, как вас зовут?'
 name = gets.chomp.capitalize
 puts 'Введите рост'
-growth = gets.to_i
+growth = gets.to_f
 ideal_weight = growth - 110
 if ideal_weight < 0
   puts "Привет #{name}, твой вес оптимальный"
